ADOP-647 change
