--- src/main/steps/review-pay-submit/payment/pay-your-fee/template.njk
@@ -1,1 +1,0 @@
-{% extends "common/template.njk" %}
